requirements.txt
pypdf
langchain
langchain-google-genai
google-generativeai
chromadb
streamlit